requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Deprecated==1.2.14
ecdsa==0.19.0
fastapi==0.110.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jose==1.0.0
limits==3.12.0
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
